[Chore] Extract bundle worker contracts and codec
