+++ crates/sekai-asset-pipeline/Cargo.toml
@@ -9,7 +9,10 @@ publish = false
 [dependencies]
 aes = "0.9"
 cbc = { version = "0.2", features = ["alloc"] }
+cridecoder = "0.3.5"
 hex = "0.4"
 rmp-serde = "1.3"
 serde = { version = "1.0", features = ["derive"] }
 thiserror = "2.0"
+tokio = { version = "1.52", features = ["macros", "rt", "time"] }
+tracing = "0.1"
