+++ crates/sekai-asset-pipeline/src/error.rs
@@ -1,4 +1,26 @@
+use std::path::PathBuf;
+
 use thiserror::Error;
+
+#[derive(Debug, Error)]
+pub enum CodecError {
+    #[error("path {0} is not valid UTF-8 for cridecoder file APIs")]
+    NonUtf8Path(PathBuf),
+    #[error("io error at {path}: {source}")]
+    Io {
+        path: PathBuf,
+        #[source]
+        source: std::io::Error,
+    },
+    #[error("ACB extraction failed: {0}")]
+    Acb(String),
+    #[error("USM extraction failed: {0}")]
+    Usm(String),
+    #[error("USM metadata read failed: {0}")]
+    Metadata(String),
+    #[error("HCA decode failed: {0}")]
+    Hca(String),
+}
 
 /// Failures produced by the reusable bundle pipeline boundary.
 #[derive(Debug, Error)]
