+++ src/core/errors.rs
@@ -2,6 +2,8 @@ use std::path::PathBuf;
 
 use thiserror::Error;
 
+pub use sekai_asset_pipeline::CodecError;
+
 /// Flatten a `reqwest::Error` together with its source chain into a single string so persisted job
 /// failure messages keep the underlying DNS/TLS/connect cause instead of only the top-level line.
 pub(crate) fn format_reqwest_error_chain(err: &reqwest::Error) -> String {
@@ -126,26 +128,6 @@ pub enum DownloadRecordError {
         #[source]
         source: opendal::Error,
     },
-}
-
-#[derive(Debug, Error)]
-pub enum CodecError {
-    #[error("path {0} is not valid UTF-8 for cridecoder file APIs")]
-    NonUtf8Path(PathBuf),
-    #[error("io error at {path}: {source}")]
-    Io {
-        path: PathBuf,
-        #[source]
-        source: std::io::Error,
-    },
-    #[error("ACB extraction failed: {0}")]
-    Acb(String),
-    #[error("USM extraction failed: {0}")]
-    Usm(String),
-    #[error("USM metadata read failed: {0}")]
-    Metadata(String),
-    #[error("HCA decode failed: {0}")]
-    Hca(String),
 }
 
 #[derive(Debug, Error)]
