+++ src/core/retry.rs
@@ -1,18 +1,6 @@
-use std::fmt::Display;
-use std::future::Future;
-use std::sync::atomic::{AtomicU64, Ordering};
-use std::time::Duration;
-
-use tokio::time::sleep;
-use tracing::warn;
+pub use sekai_asset_pipeline::{retry_async, retry_sync, RetryPolicy};
 
 use crate::core::config::RetryConfig;
-
-pub trait RetryPolicy {
-    fn attempts(&self) -> usize;
-    fn initial_backoff_ms(&self) -> u64;
-    fn max_backoff_ms(&self) -> u64;
-}
 
 impl RetryPolicy for RetryConfig {
     fn attempts(&self) -> usize {
@@ -28,127 +16,6 @@ impl RetryPolicy for RetryConfig {
     }
 }
 
-impl RetryPolicy for sekai_asset_pipeline::RetryOptions {
-    fn attempts(&self) -> usize {
-        self.attempts
-    }
-
-    fn initial_backoff_ms(&self) -> u64 {
-        self.initial_backoff_ms
-    }
-
-    fn max_backoff_ms(&self) -> u64 {
-        self.max_backoff_ms
-    }
-}
-
-pub async fn retry_async<T, E, Op, Fut, ShouldRetry, Policy>(
-    config: &Policy,
-    operation: &str,
-    mut op: Op,
-    should_retry: ShouldRetry,
-) -> Result<T, E>
-where
-    E: Display,
-    Op: FnMut(usize) -> Fut,
-    Fut: Future<Output = Result<T, E>>,
-    ShouldRetry: Fn(&E) -> bool,
-    Policy: RetryPolicy,
-{
-    let attempts = config.attempts().max(1);
-    for attempt in 1..=attempts {
-        match op(attempt).await {
-            Ok(value) => return Ok(value),
-            Err(err) if attempt < attempts && should_retry(&err) => {
-                let delay = backoff_delay(config, attempt);
-                warn!(
-                    operation,
-                    attempt,
-                    max_attempts = attempts,
-                    delay_ms = delay.as_millis(),
-                    error = %err,
-                    "operation failed, retrying"
-                );
-                sleep(delay).await;
-            }
-            Err(err) => return Err(err),
-        }
-    }
-
-    unreachable!("retry_async must return from within the attempt loop")
-}
-
-pub fn retry_sync<T, E, Op, ShouldRetry, Policy>(
-    config: &Policy,
-    operation: &str,
-    mut op: Op,
-    should_retry: ShouldRetry,
-) -> Result<T, E>
-where
-    E: Display,
-    Op: FnMut(usize) -> Result<T, E>,
-    ShouldRetry: Fn(&E) -> bool,
-    Policy: RetryPolicy,
-{
-    let attempts = config.attempts().max(1);
-    for attempt in 1..=attempts {
-        match op(attempt) {
-            Ok(value) => return Ok(value),
-            Err(err) if attempt < attempts && should_retry(&err) => {
-                let delay = backoff_delay(config, attempt);
-                warn!(
-                    operation,
-                    attempt,
-                    max_attempts = attempts,
-                    delay_ms = delay.as_millis(),
-                    error = %err,
-                    "operation failed, retrying"
-                );
-                std::thread::sleep(delay);
-            }
-            Err(err) => return Err(err),
-        }
-    }
-
-    unreachable!("retry_sync must return from within the attempt loop")
-}
-
-fn backoff_delay(config: &impl RetryPolicy, attempt: usize) -> Duration {
-    let base = config.initial_backoff_ms().max(1);
-    let max = config.max_backoff_ms().max(base);
-    let multiplier = 1u64
-        .checked_shl((attempt.saturating_sub(1)) as u32)
-        .unwrap_or(u64::MAX);
-    let capped = base.saturating_mul(multiplier).min(max);
-    // "Equal jitter": keep half the computed delay fixed and randomize the other half so that many
-    // concurrent operations sharing one RetryConfig don't retry in lockstep (thundering herd).
-    let half = capped / 2;
-    let jitter = if half > 0 {
-        jitter_noise() % (half + 1)
-    } else {
-        0
-    };
-    Duration::from_millis(half.saturating_add(jitter))
-}
-
-/// Cheap, dependency-free per-call noise for backoff jitter. Mixes a process-global counter with
-/// the wall-clock subsecond nanos through a splitmix64 finalizer; not cryptographic, but enough to
-/// decorrelate concurrent retries.
-fn jitter_noise() -> u64 {
-    static COUNTER: AtomicU64 = AtomicU64::new(0);
-    let nanos = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map(|d| d.subsec_nanos() as u64)
-        .unwrap_or(0);
-    let mut x = COUNTER
-        .fetch_add(1, Ordering::Relaxed)
-        .wrapping_mul(0x9E37_79B9_7F4A_7C15)
-        ^ nanos;
-    x = (x ^ (x >> 30)).wrapping_mul(0xBF58_476D_1CE4_E5B9);
-    x = (x ^ (x >> 27)).wrapping_mul(0x94D0_49BB_1331_11EB);
-    x ^ (x >> 31)
-}
-
 #[cfg(test)]
 mod tests {
     use std::sync::atomic::{AtomicUsize, Ordering};
@@ -157,7 +24,7 @@ mod tests {
     use crate::core::config::RetryConfig;
 
     #[tokio::test]
-    async fn retry_async_retries_until_success() {
+    async fn retry_async_accepts_the_application_retry_config() {
         let attempts = AtomicUsize::new(0);
         let config = RetryConfig {
             attempts: 3,
@@ -186,7 +53,7 @@ mod tests {
     }
 
     #[test]
-    fn retry_sync_stops_on_non_retryable_error() {
+    fn retry_sync_accepts_the_application_retry_config() {
         let attempts = AtomicUsize::new(0);
         let config = RetryConfig {
             attempts: 4,
@@ -194,7 +61,7 @@ mod tests {
             max_backoff_ms: 1,
         };
 
-        let err = retry_sync(
+        let error = retry_sync(
             &config,
             "test sync",
             |_| {
@@ -205,7 +72,7 @@ mod tests {
         )
         .unwrap_err();
 
-        assert_eq!(err, "fatal");
+        assert_eq!(error, "fatal");
         assert_eq!(attempts.load(Ordering::SeqCst), 1);
     }
 }
